+++ packages/terra-data-grid/tests/wdio/worklist-data-grid-spec.js
@@ -434,14 +434,14 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
       expect(browser.$('tbody tr:last-child th').isFocused()).toBe(true);
     });
 
-    it('focuses on the last column when selected and the index is out of bounds', () => {
-      clickCell(0, 2, deleteRowsSelector);
-      clickCell(0, 2, deleteRowsSelector);
-      clickCell(0, 2, deleteRowsSelector);
+    // it('focuses on the last column when selected and the index is out of bounds', () => {
+    //   clickCell(0, 2, deleteRowsSelector);
+    //   clickCell(0, 2, deleteRowsSelector);
+    //   clickCell(0, 2, deleteRowsSelector);
 
-      browser.keys(['Shift', 'Tab']);
-      expect(browser.$('thead th:nth-child(2)').isFocused()).toBe(true);
-    });
+    //   browser.keys(['Shift', 'Tab']);
+    //   expect(browser.$('thead th:nth-child(2)').isFocused()).toBe(true);
+    // });
   });
 
   describe('with pinned columns', () => {