+++ packages/terra-table/CHANGELOG.md
@@ -11,6 +11,7 @@
   
 * Changed
   * Removed console warning message when no pinned columns exist.
+  * Updated the table component to properly support selection callbacks for the grid context.
 
 ## 5.2.1 - (December 1, 2023)